Flash support
To enable flash we need to set the commandline switch 'ppapi-flash-path' for the electron app.
We look for the file .pennywise-flash in the users home dir expecting to be a copy of a the PepperflashPlugin or a symlink to it.
A users flash install can be found in chrome://flash.